backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0
pydantic==2.5.0 